[Bugfix] Use int64_t for memory budget

+++ src/pass/rematerialization.cc
@@ -1001,7 +1001,7 @@ TVM_REGISTER_PASS_CONFIG_OPTION("raf.remat.use_gflops_cost", IntImm);
 
 Pass Rematerialization() {
   PassContext pass_ctx = PassContext::Current();
-  int memory_budget =
+  int64_t memory_budget =
       pass_ctx->GetConfig("raf.memory_budget", Integer(static_cast<int>(0))).value().IntValue();
   // Turn profiler on by default. With caching it is pretty fast now.
   bool use_profiler = !(pass_ctx->GetConfig("raf.remat.use_gflops_cost", Bool(false)).value());
